+++ frontend/src/pages/ReportPage.js
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useNavigate } from 'react-router-dom';
-import _ from 'lodash';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
@@ -13,106 +12,34 @@ import TabContext from '@mui/lab/TabContext';
 import TabList from '@mui/lab/TabList';
 import TabPanel from '@mui/lab/TabPanel';
 
-import { red } from '@mui/material/colors';
-
-// Sweetalert
-import Swal from 'sweetalert2';
-import withReactContent from 'sweetalert2-react-content';
-
-// components
-import Iconify from '../components/iconify';
-
 // Lib Auth
 import { verifyToken, getTokenFromLocalStorage } from '../libs/Auth';
 
-// Context
-import { useAuth } from '../contexts/AuthContext';
-
 // Section Report
 import ReportSummary1 from '../sections/reports/ReportSummary1';
-// import ReportSummary2 from '../sections/reports/ReportSummary2';
 import ReportSummary3 from '../sections/reports/ReportSummary3';
-// import ReportSummary4 from '../sections/reports/ReportSummary4';
-import ReportSummary5 from '../sections/reports/ReportSummary5';
-// import ReportSummary6 from '../sections/reports/ReportSummary6';
 import ReportSummary7 from '../sections/reports/ReportSummary7';
 import ReportSummary8 from '../sections/reports/ReportSummary8';
 
-const color = red[500];
-
-// Notify Toast Config
-const MySwal = withReactContent(Swal);
-const Toast = MySwal.mixin({
-  toast: true,
-  position: 'center',
-  showConfirmButton: false,
-  timer: 3000,
-  timerProgressBar: true,
-  didOpen: (toast) => {
-    toast.addEventListener('mouseenter', MySwal.stopTimer);
-    toast.addEventListener('mouseleave', MySwal.resumeTimer);
-  },
-});
-// Notify Toast Config
-
 export default function ReportPage() {
-  const auth = useAuth();
   const navigate = useNavigate();
-  const [token, setToken] = useState(auth.accessToken);
-  const [open, setOpen] = useState(null);
-
   const [tabValue, setTabValue] = useState('1');
 
   const handleChange = (event, newValue) => {
     setTabValue(newValue);
   };
 
-  const handleCatchAxios = (errorCatch, sec) => {
-    if (errorCatch.response) {
-      const { status } = errorCatch.response;
-      if (status === 404) {
-        setTimeout(() => {
-          Toast.fire({
-            icon: 'error',
-            title: 'ไม่พบข้อมูล',
-          });
-        }, 1000);
-      } else {
-        Toast.fire({
-          icon: 'error',
-          title: 'เกิดข้อผิดพลาดบางอย่าง',
-        });
-      }
-    } else if (errorCatch.request) {
-      Toast.fire({
-        icon: 'error',
-        title: 'เกิดข้อผิดพลาดบางอย่าง, ขออภัยในความไม่สะดวก!',
-      });
-    } else {
-      Toast.fire({
-        icon: 'error',
-        title: 'เกิดข้อผิดพลาดบางอย่าง, ขออภัยในความไม่สะดวก!',
-      });
-    }
-    setOpen(null);
-  };
-
   useEffect(() => {
     async function checkVerifyToken() {
       const auth_token = getTokenFromLocalStorage('access_token');
       const verify = await verifyToken(auth_token);
-      const { statusCode, profile, access_token } = verify;
-      if (statusCode === 200 && profile) {
-        if (access_token) {
-          setToken(access_token);
-        } else {
-          navigate('/login', { replace: true });
-        }
-      } else {
+      const { statusCode, profile, access_token } = verify || {};
+      if (!(statusCode === 200 && profile && access_token)) {
         navigate('/login', { replace: true });
       }
     }
     checkVerifyToken();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (